+++ functions/functions-backend.php
@@ -3,7 +3,7 @@
  * Backend related functions
  *
  * @author      Flurin Dürst
- * @version     1.0
+ * @version     1.1
  * @since       WPSeed 0.11.3
  *
  */
@@ -64,6 +64,18 @@
     return $mimes;
   }
   add_filter('upload_mimes', 'cc_mime_types');
+
+/* MISC
+/===================================================== */
+
+  /* Hide core update info for non-admins
+  /------------------------*/
+  function onlyadmin_update() {
+    if (!current_user_can('update_core')) {
+        remove_action( 'admin_notices', 'update_nag', 3 );
+    }
+  }
+  add_action( 'admin_head', 'honlyadmin_update', 1 );
 
 
 ?>